version 2 for loading materials on run time
in this code it used to load material texture on run time by accessing url (string ) it used to optimized project ,material have to attached with certain objects added in array and required url put on same index as material puts on. to use this script in unity all you need to do add mesh  render by click on add component. on inspector bar.